Fix occasional, spurious test failure in 'csv_col_to_sparse_dims' test
In this test we create a duplicate value on the dimension. However, the duplicatate
can be stored/retrieved in arbitrary order. When the order is different
than TileDB retrieval order is different than the original sort, the comparison
against original dataframe fails due to ordering even though data is correct.